requirements.txt
pydantic==2.4.2
python-dotenv==1.0.0
textX[cli]==3.1.1
streamlit==1.27.2
streamlit-chat==0.1.1
openai~=1.30.1
msal==1.24.1
Office365-REST-Python-Client==2.4.4